Split IssueTest into individual tests
