+++ tests/Issue/Issue87Test.php
@@ -1,0 +1,71 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SimpleThings\EntityAudit\Tests\Issue;
+
+use SimpleThings\EntityAudit\Tests\BaseTest;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\Issue87AbstractProject;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\Issue87Organization;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\Issue87Project;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\Issue87ProjectComment;
+
+final class Issue87Test extends BaseTest
+{
+    protected $schemaEntities = [
+        Issue87Project::class,
+        Issue87ProjectComment::class,
+        Issue87AbstractProject::class,
+        Issue87Organization::class,
+    ];
+
+    protected $auditedEntities = [
+        Issue87Project::class,
+        Issue87ProjectComment::class,
+        Issue87AbstractProject::class,
+        Issue87Organization::class,
+    ];
+
+    public function testIssue87(): void
+    {
+        $org = new Issue87Organization();
+        $project = new Issue87Project();
+        $project->setOrganisation($org);
+        $project->setSomeProperty('some property');
+        $project->setTitle('test project');
+        $comment = new Issue87ProjectComment();
+        $comment->setProject($project);
+        $comment->setText('text comment');
+
+        $this->em->persist($org);
+        $this->em->persist($project);
+        $this->em->persist($comment);
+        $this->em->flush();
+
+        $auditReader = $this->auditManager->createAuditReader($this->em);
+
+        $auditedProject = $auditReader->find(\get_class($project), $project->getId(), 1);
+
+        static::assertSame($org->getId(), $auditedProject->getOrganisation()->getId());
+        static::assertSame('test project', $auditedProject->getTitle());
+        static::assertSame('some property', $auditedProject->getSomeProperty());
+
+        $auditedComment = $auditReader->find(\get_class($comment), $comment->getId(), 1);
+        static::assertSame('test project', $auditedComment->getProject()->getTitle());
+
+        $project->setTitle('changed project title');
+        $this->em->flush();
+
+        $auditedComment = $auditReader->find(\get_class($comment), $comment->getId(), 2);
+        static::assertSame('changed project title', $auditedComment->getProject()->getTitle());
+    }
+}
