+++ tests/Issue/IssueDuplicateRevisionTest.php
@@ -1,0 +1,63 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SimpleThings\EntityAudit\Tests\Issue;
+
+use SimpleThings\EntityAudit\Tests\BaseTest;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\DuplicateRevisionFailureTestOwnedElement;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\DuplicateRevisionFailureTestPrimaryOwner;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\DuplicateRevisionFailureTestSecondaryOwner;
+
+final class IssueDuplicateRevisionTest extends BaseTest
+{
+    protected $schemaEntities = [
+        DuplicateRevisionFailureTestPrimaryOwner::class,
+        DuplicateRevisionFailureTestSecondaryOwner::class,
+        DuplicateRevisionFailureTestOwnedElement::class,
+    ];
+
+    protected $auditedEntities = [
+        DuplicateRevisionFailureTestPrimaryOwner::class,
+        DuplicateRevisionFailureTestSecondaryOwner::class,
+        DuplicateRevisionFailureTestOwnedElement::class,
+    ];
+
+    /**
+     * @doesNotPerformAssertions
+     */
+    public function testDuplicateRevisionKeyConstraintFailure(): void
+    {
+        $primaryOwner = new DuplicateRevisionFailureTestPrimaryOwner();
+        $this->em->persist($primaryOwner);
+
+        $secondaryOwner = new DuplicateRevisionFailureTestSecondaryOwner();
+        $this->em->persist($secondaryOwner);
+
+        $primaryOwner->addSecondaryOwner($secondaryOwner);
+
+        $element = new DuplicateRevisionFailureTestOwnedElement();
+        $this->em->persist($element);
+
+        $primaryOwner->addElement($element);
+        $secondaryOwner->addElement($element);
+
+        $this->em->flush();
+
+        $this->em->getUnitOfWork()->clear();
+
+        $primaryOwner = $this->em->find(DuplicateRevisionFailureTestPrimaryOwner::class, 1);
+
+        $this->em->remove($primaryOwner);
+        $this->em->flush();
+    }
+}
