+++ tests/Issue/IssueEscapedColumnsTest.php
@@ -1,0 +1,44 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace SimpleThings\EntityAudit\Tests\Issue;
+
+use SimpleThings\EntityAudit\Tests\BaseTest;
+use SimpleThings\EntityAudit\Tests\Fixtures\Issue\EscapedColumnsEntity;
+
+final class IssueEscapedColumnsTest extends BaseTest
+{
+    protected $schemaEntities = [
+        EscapedColumnsEntity::class,
+    ];
+
+    protected $auditedEntities = [
+        EscapedColumnsEntity::class,
+    ];
+
+    /**
+     * @doesNotPerformAssertions
+     */
+    public function testEscapedColumns(): void
+    {
+        $e = new EscapedColumnsEntity();
+        $e->setLeft(1);
+        $e->setLft(2);
+        $this->em->persist($e);
+        $this->em->flush();
+
+        $reader = $this->auditManager->createAuditReader($this->em);
+
+        $reader->find(\get_class($e), $e->getId(), 1);
+    }
+}
